requirements.txt
twisted>=18.9.0
scrapy>=1.6.0
pymysql>=0.9.0
urllib3>=1.24.1
pyexecjs>=1.5.1
openpyxl>=2.5.10
xlrd>=1.2.0
pyecharts>=0.5.11
pyecharts-snapshot>=0.2.0
jieba>=0.39
wordcloud>=1.5.0